backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-multipart==0.0.6
pdfplumber==0.11.0
python-docx==1.1.0